+++ CHANGELOG.md
@@ -1,5 +1,79 @@
 # Change Log
 
-## 2017-01-23 Chase Hennion
+## 2017-2-19 Chase Hennion
+
+* Cleaning up BuildingEscape code
+* Adding comments to each class
+* Adding custom wall material
+* Adding SlidingWall blueprint
+* Cleaning up lighting
+* Game is ready to have an initial release
+
+## 2017-2-14 Chase Hennion
+
+* Added input component for throwing
+* Finished sliding door logic
+* Finished throwing puzzle logic
+* Cleaned up game world
+* Cleaned up brick wall stretched texture
+* Game is now complete
+* Will now move on to cleaning up code
+
+## 2017-2-13 Chase Hennion
+
+* Added throw logic to grabber. Still need to add input component to trigger it.
+* Added last three rooms. Still need to add trigger volumes to plates.
+* Added toggle to opendoor to allow for one time door opening when a trigger volume is tripped.
+
+## 2017-2-11 Chase Hennion
+
+* Added Multiplate Puzzle actor component
+* Updated second room to use the multiplate puzzle
+
+## 2017-2-9 Chase Hennion
+
+* Adding work that covers lectures 90 through 94 in Section 3.
+* Section 3 lectures are now complete, moving on to cleaning up and improving the game.
+
+## 2017-2-7 Chase Hennion
+
+* Adding work that covers lectures 87 through 89 in Section 3.
+
+## 2017-2-6 Chase Hennion
+
+* Adding work that covers lectures 84 through 86 in Section 3.
+
+## 2017-2-5 Chase Hennion
+
+* Adding work that covers lectures 82 and 83 in Section 3.
+
+## 2017-2-4 Chase Hennion
+
+* Adding work that covers lectures 77 through 81 in Section 3.
+
+## 2017-2-2 Chase Hennion
+
+* Adding work that covers lectures 74 through 77 in Section 3.
+
+## 2017-1-31 Chase Hennion
+
+* Adding work that covers lectures 70 through 73 in Section 3.
+
+## 2017-1-29 Chase Hennion
+
+* Adding work that covers lectures 63 through 66 in Section 3.
+* Adding work that covers lectures 67 through 69 in Section 3.
+
+## 2017-1-28 Chase Hennion
+
+* Adding work that covers lectures 58 through 62 in Section 3
+
+## 2017-01-26 Chase Hennion
+
+* Adding work that covers lectures 55 through 57 in Section 3.
+
+## 2017-01-24 Chase Hennion
 
 * Initial commit of boilerplate Visual Studio and Unreal content.
+* Adding initial readme.
+* Work covers lectures 51 through 54 in Section 3.